# backend/src/main/java/com/delta/taskmanagement/dto/TaskUpdateRequest.java
package com.delta.taskmanagement.dto;

import com.delta.taskmanagement.enums.Priority;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class TaskUpdateRequest {

    @NotBlank
    @Size(max = 255)
    private String title;

    private String description;

    @NotNull
    private Priority priority;

    @NotNull
    private Long version;
}